Move PackageDetailResource tests to new test case.
